refactor: split form and dialog helpers out of large mixins
Extract form state/helpers and confirm dialogs; trim layout,
persistence, and execution mixins and wire them in main_window MRO.
