+++ gui/mixins/__init__.py
@@ -1,8 +1,11 @@
 from gui.browse_dialogs import BrowseDialogsMixin
 from gui.mixins.binary_ui import BinaryUIMixin
+from gui.mixins.confirm_dialog import ConfirmDialogMixin
 from gui.mixins.connection import ConnectionMixin
 from gui.mixins.diagnostics import DiagnosticsMixin
 from gui.mixins.execution import ExecutionMixin
+from gui.mixins.form_helpers import FormHelpersMixin
+from gui.mixins.form_state import FormStateMixin
 from gui.mixins.layout import LayoutMixin
 from gui.mixins.menu import MenuMixin
 from gui.mixins.persistence import PersistenceMixin
@@ -13,9 +16,12 @@ from gui.mixins.theme_mixin import ThemeMixin
 __all__ = [
     "BinaryUIMixin",
     "BrowseDialogsMixin",
+    "ConfirmDialogMixin",
     "ConnectionMixin",
     "DiagnosticsMixin",
     "ExecutionMixin",
+    "FormHelpersMixin",
+    "FormStateMixin",
     "LayoutMixin",
     "MenuMixin",
     "PersistenceMixin",
